requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.6
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
discord==1.0.1
discord.py==1.6.0
dnspython==2.1.0
idna==2.10
multidict==5.1.0
Pillow==8.2.0
pycodestyle==2.7.0
pymongo==3.5.1
python-dotenv==0.15.0
requests==2.25.1
soupsieve==2.2.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
yarl==1.6.3
